+++ scripts/cl.sh
@@ -395,10 +395,20 @@ main()
 		execution_slug="$2"
 
 		execution_path="$(execution_derive_path_from_slug "$executions" "$execution_slug")"
+		execution_bin="$(mktemp --tmpdir="$(dirname "${execution_path}")")"
+		trap 'rm $execution_bin' EXIT
+
 		sed --quiet -E "/[\`]{3}([a-z]+ )?name=${execution_script}/,/[\`]{3}/p" "${execution_path}" |
-			sed '1d;$d' |
-			( SHELL=/usr/bin/sh sh 2>&1 || kill $$ ) |
-			"$script" attach output "${execution_slug}" "${execution_script}"
+			sed '1d;$d' > "$execution_bin"
+		chmod +x "$execution_bin"
+		if head -n1 "$execution_bin" | grep -q '^#!';
+		then
+			( SHELL=/usr/bin/sh "$execution_bin" 2>&1 || kill $$ ) |
+				"$script" attach output "${execution_slug}" "${execution_script}"
+		else
+			( SHELL=/usr/bin/sh sh "$execution_bin" 2>&1 || kill $$ ) |
+				"$script" attach output "${execution_slug}" "${execution_script}"
+		fi
 		if grep -E "[\`]{3}([a-z]+ )?name=${execution_script}" "$execution_path" | grep -q notify
 		then
 			run_notify "${execution_script}"