+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ## [Unreleased]
 
 1. Improve precision of cleanups to support concurrent checklist execution
+2. Allow embedding and execution of abitrary scripts in checklists
+   - See the documentation for `cl run ...`